pl/math: Add scalar asinh
The new routine uses a similar approach to asinhf, using a polynomial
only in the region where either returning x or calculating the result
directly is not sufficiently precise.

Worst-case error is about 2 ULP, close to 1. There are 4
intervals with slightly different error behaviour, as follows:

Interval              Worst-case accuracy (ulp)
|x| < 2^-26           0.0
|x| < 1               1.5
|x| < ~sqrt(DBL_MAX)  2.0
|x| < infinity        1.0

log has been copied from the main math directory so that it can be
used in asinh. The only modifications to the relevant files are to
remove aliases and rename log itself to an internal 'helper' name.